refactor: 블럭 삭제 API DDD + hexagonal 로 리팩토링
