--- src/main/java/com/joojoo/api/block/application/BlockService.java
@@ -1,8 +1,0 @@
-package com.joojoo.api.block.application;
-
-import com.joojoo.api.block.domain.model.enums.DeleteMode;
-
-public interface BlockService {
-
-    void deleteBlock(Long userId, Long blockId, DeleteMode mode);
-}
