+++ src/main/java/com/joojoo/api/block/application/port/in/DeleteBlockUseCase.java
@@ -1,5 +1,9 @@
 package com.joojoo.api.block.application.port.in;
 
+import com.joojoo.api.block.domain.model.enums.DeleteMode;
+
 public interface DeleteBlockUseCase {
     void deleteAllBlockMappings(Long userId);
+
+    void deleteBlock(Long userId, Long blockId, DeleteMode mode);
 }
